requirements.txt
numpy>=1.16.0
scipy>=1.2.0
pillow>=5.2.0
